requirement.txt
pandas
sqlalchemy
python-dotenv
psycopg2
psycopg2-binary
matplotlib
seaborn
